Fixing warnings.

+++ Code/Store3Imap/ScribeImap_Mail.cpp
@@ -241,7 +241,7 @@ void ImapMail::Load()
 					break;
 				}
 
-				if (Buf.Length() >= f.GetSize())
+				if ((int64)Buf.Length() >= f.GetSize())
 					break;
 			}
